fix compile error for bfloat16

+++ oneflow/user/kernels/fused_relu_dropout_grad_kernel.cu
@@ -50,7 +50,8 @@ __global__ void VectorizedReluDropoutBitmaskBackwardKernel(
       const int32_t col_mod_warpsize = col % kWarpSize;
       const IndexType aux_idx = ((row * aux_ld) + col) / kWarpSize;
       bool is_positive = mask[aux_idx] & (1 << col_mod_warpsize);
-      dx_vec.elem[i] = dy_vec.elem[i] * static_cast<T>(is_positive) * static_cast<T>(scale);
+      dx_vec.elem[i] =
+          dy_vec.elem[i] * static_cast<T>(static_cast<float>(is_positive)) * static_cast<T>(scale);
     }
     *(reinterpret_cast<LoadStoreType*>(dx + linear_pack_index)) = dx_vec.storage;
   }
@@ -62,7 +63,8 @@ __global__ void VectorizedReluDropoutBitmaskBackwardKernel(
     const IndexType tail_col_mod_warpsize = tail_col % kWarpSize;
     const IndexType tail_aux_idx = ((tail_row * aux_ld) + tail_col) / kWarpSize;
     bool is_positive = mask[tail_aux_idx] & (1 << tail_col_mod_warpsize);
-    dx[tail_index] = dy[tail_index] * static_cast<T>(is_positive) * static_cast<T>(scale);
+    dx[tail_index] =
+        dy[tail_index] * static_cast<T>(static_cast<float>(is_positive)) * static_cast<T>(scale);
   }
 }
